Replace repetitive Decimal conversion logic with reusable utility
Consolidates and centralizes the logic for converting Decimal objects
to floats in a new `convert_decimals` utility function. Removes
duplicate methods from multiple classes and updates all relevant
call sites to use the new utility. This improves code maintainability
and reduces redundancy.

Additionally, includes minor logging adjustments, test updates to
account for the new utility, and fixes event loop issues in test
fixtures.
