+++ src/v4vapp_backend_v2/process/process_custom_json.py
@@ -1,4 +1,3 @@
-import asyncio
 from datetime import datetime, timezone
 from typing import List
 
@@ -6,21 +5,124 @@ from colorama import Fore, Style
 
 from v4vapp_backend_v2.accounting.account_balances import keepsats_balance
 from v4vapp_backend_v2.accounting.ledger_account_classes import LiabilityAccount, RevenueAccount
-from v4vapp_backend_v2.accounting.ledger_entry_class import LedgerEntry, LedgerType
+from v4vapp_backend_v2.accounting.ledger_entry_class import (
+    LedgerEntry,
+    LedgerEntryCreationException,
+    LedgerType,
+)
 from v4vapp_backend_v2.actions.tracked_any import load_tracked_object
-from v4vapp_backend_v2.actions.tracked_models import TrackedBaseModel
+from v4vapp_backend_v2.actions.tracked_models import ReplyType, TrackedBaseModel
 from v4vapp_backend_v2.config.setup import InternalConfig, logger
 from v4vapp_backend_v2.helpers.crypto_conversion import CryptoConversion
 from v4vapp_backend_v2.helpers.currency_class import Currency
 from v4vapp_backend_v2.hive_models.custom_json_data import KeepsatsTransfer
 from v4vapp_backend_v2.hive_models.op_custom_json import CustomJson
 from v4vapp_backend_v2.hive_models.return_details_class import HiveReturnDetails, ReturnAction
+from v4vapp_backend_v2.models.invoice_models import Invoice
 from v4vapp_backend_v2.process.hive_notification import reply_with_hive
 from v4vapp_backend_v2.process.hold_release_keepsats import release_keepsats
 from v4vapp_backend_v2.process.process_errors import (
     CustomJsonToLightningError,
     InsufficientBalanceError,
 )
+from v4vapp_backend_v2.process.process_invoice import process_lightning_receipt_stage_2
+from v4vapp_backend_v2.process.process_transfer import follow_on_transfer
+
+
+# MARK: CustomJson Operations
+async def process_custom_json(
+    custom_json: CustomJson, nobroadcast: bool = False
+) -> List[LedgerEntry]:
+    """
+    Processes a CustomJson operation and creates a ledger entry if applicable.
+
+    This method handles CustomJson operations, ensuring that appropriate debit and credit
+    accounts are assigned based on the operation type. If a ledger entry with the same group_id
+    already exists, the operation is skipped.
+
+    Returns:
+        LedgerEntry: The created or existing ledger entry, or None if no entry is created.
+    """
+    server_id = InternalConfig().server_id
+    if custom_json.cj_id in ["v4vapp_notification", "v4vapp_dev_notification"]:
+        logger.info(f"Notification CustomJson: {custom_json.json_data.memo}")
+        return None
+    if custom_json.cj_id in ["v4vapp_dev_transfer", "v4vapp_transfer"]:
+        keepsats_transfer = KeepsatsTransfer.model_validate(custom_json.json_data)
+        keepsats_transfer.msats = (
+            (keepsats_transfer.sats * 1000)
+            if keepsats_transfer.sats and not keepsats_transfer.msats
+            else keepsats_transfer.msats
+        )
+        # MARK: CustomJson Transfer user to user
+        if (
+            custom_json.from_account
+            and custom_json.to_account
+            and keepsats_transfer.msats
+            and custom_json.from_account != custom_json.to_account
+        ):
+            if not custom_json.conv or custom_json.conv.is_unset():
+                quote = await TrackedBaseModel.nearest_quote(timestamp=custom_json.timestamp)
+                await custom_json.update_conv(quote=quote)
+            ledger_entries = await custom_json_internal_transfer(
+                custom_json=custom_json, keepsats_transfer=keepsats_transfer
+            )
+            # Check for a parent id to see if this is a reply transaction
+            if keepsats_transfer.parent_id:
+                # This is a reply transaction, we need to process it as such
+
+                parent_op = await load_tracked_object(tracked_obj=keepsats_transfer.parent_id)
+                if parent_op:
+                    parent_op.add_reply(
+                        reply_id=custom_json.group_id_p,
+                        reply_type=ReplyType.CUSTOM_JSON,
+                        reply_msat=keepsats_transfer.msats if keepsats_transfer.msats else 0,
+                        reply_message="Reply to transfer",
+                    )
+                    await parent_op.save()
+                    # Now we may need to process if the parent op is an Invoice we need to send the keepsats
+                    # to the correct destination.
+                    if isinstance(parent_op, Invoice) and custom_json.from_account == server_id:
+                        await process_lightning_receipt_stage_2(
+                            invoice=parent_op, nobroadcast=nobroadcast
+                        )
+                        return ledger_entries
+                    if isinstance(parent_op, CustomJson) and custom_json.to_account == server_id:
+                        # Process this as if it were a request to convert Keepsats to Hive/HBD
+                        logger.info(
+                            f"Processing Keepsats to Hive conversion: {custom_json.json_data.memo}"
+                        )
+                        # await conversion_keepsats_to_hive(server_id= server_id, cust_id=tracked_op=custom_json, nobroadcast=nobroadcast)
+
+            if custom_json.to_account == server_id:
+                # Process this as if it were an inbound Hive transfer with a memo.
+                await follow_on_transfer(tracked_op=custom_json, nobroadcast=nobroadcast)
+
+            return []
+
+        # MARK: CustomJson to pay a lightning invoice
+        # If this has a memo that should contain the invoice and the instructions like "#clean"
+        # invoice_message we will use to send on if we generate an invoice form a lightning address
+        elif keepsats_transfer.memo and keepsats_transfer.to_account == server_id:
+            # This is a transfer operation, we need to process it as such
+
+            if not custom_json.conv or custom_json.conv.is_unset():
+                await custom_json.update_conv()
+                if custom_json.conv and custom_json.conv.is_unset():
+                    raise LedgerEntryCreationException(
+                        "Conversion not set in CustomJson operation."
+                    )
+
+            await follow_on_transfer(tracked_op=custom_json, nobroadcast=nobroadcast)
+            return []
+
+    logger.error(
+        f"CustomJson operation not implemented for v4vapp_group_id: {custom_json.group_id}.",
+        extra={"notification": False, **custom_json.log_extra},
+    )
+    raise NotImplementedError(
+        f"Some other custom_json functionality which hasn't been implemented yet {custom_json.group_id}."
+    )
 
 
 async def custom_json_internal_transfer(
