+++ src/v4vapp_backend_v2/process/process_hive.py
@@ -1,5 +1,5 @@
 from datetime import datetime, timezone
-from typing import List, Union
+from typing import List
 
 from v4vapp_backend_v2.accounting.ledger_account_classes import AssetAccount, LiabilityAccount
 from v4vapp_backend_v2.accounting.ledger_entry_class import (
@@ -9,20 +9,17 @@ from v4vapp_backend_v2.accounting.ledger_entry_class import (
     LedgerEntryException,
     LedgerType,
 )
-from v4vapp_backend_v2.actions.tracked_any import TrackedAny, TrackedTransfer, load_tracked_object
-from v4vapp_backend_v2.actions.tracked_models import ReplyType, TrackedBaseModel
+from v4vapp_backend_v2.actions.tracked_any import TrackedAny, TrackedTransfer
+from v4vapp_backend_v2.actions.tracked_models import ReplyType
 from v4vapp_backend_v2.config.setup import InternalConfig, logger
-from v4vapp_backend_v2.helpers.crypto_conversion import CryptoConv
 from v4vapp_backend_v2.helpers.general_purpose_funcs import lightning_memo
-from v4vapp_backend_v2.hive_models.custom_json_data import KeepsatsTransfer
 from v4vapp_backend_v2.hive_models.op_custom_json import CustomJson
 from v4vapp_backend_v2.hive_models.op_fill_order import FillOrder
 from v4vapp_backend_v2.hive_models.op_limit_order_create import LimitOrderCreate
 from v4vapp_backend_v2.hive_models.op_transfer import TransferBase
-from v4vapp_backend_v2.models.invoice_models import Invoice
-from v4vapp_backend_v2.process.process_custom_json import custom_json_internal_transfer
+from v4vapp_backend_v2.process.process_custom_json import process_custom_json
 from v4vapp_backend_v2.process.process_errors import HiveLightningError
-from v4vapp_backend_v2.process.process_invoice import process_lightning_receipt_stage_2
+from v4vapp_backend_v2.process.process_orders import process_create_fill_order_op
 from v4vapp_backend_v2.process.process_resend_hive import resend_transactions
 from v4vapp_backend_v2.process.process_transfer import HiveTransferError, follow_on_transfer
 
@@ -264,250 +261,3 @@ async def process_transfer_op(
             logger.exception(f"Follow-on task failed: {e}", extra={"notification": False})
 
     return ledger_entry
-
-
-async def process_create_fill_order_op(
-    limit_fill_order: Union[LimitOrderCreate, FillOrder], nobroadcast: bool = False
-) -> List[LedgerEntry]:
-    ledger_entries = []
-    tracked_customers = [InternalConfig().server_id]  # Replace with dynamic query if needed
-
-    if isinstance(limit_fill_order, LimitOrderCreate):
-        logger.info(f"Limit order create: {limit_fill_order.orderid}")
-        if not limit_fill_order.conv or limit_fill_order.conv.is_unset():
-            quote = await TrackedBaseModel.nearest_quote(timestamp=limit_fill_order.timestamp)
-            limit_fill_order.conv = CryptoConv(
-                conv_from=limit_fill_order.amount_to_sell.unit,  # HIVE
-                value=limit_fill_order.amount_to_sell.amount_decimal,  # 25.052 HIVE
-                converted_value=limit_fill_order.min_to_receive.amount_decimal,  # 6.738 HBD
-                quote=quote,
-                timestamp=limit_fill_order.timestamp,
-            )
-        ledger_entry = LedgerEntry(
-            group_id=limit_fill_order.group_id,
-            short_id=limit_fill_order.short_id,
-            timestamp=limit_fill_order.timestamp,
-            op_type=limit_fill_order.op_type,
-        )
-        ledger_entry.debit = AssetAccount(name="Escrow Hive", sub=limit_fill_order.owner)
-        ledger_entry.credit = AssetAccount(
-            name="Customer Deposits Hive", sub=limit_fill_order.owner
-        )
-        ledger_entry.description = limit_fill_order.ledger_str
-        ledger_entry.ledger_type = LedgerType.LIMIT_ORDER_CREATE
-        ledger_entry.debit_unit = ledger_entry.credit_unit = limit_fill_order.amount_to_sell.unit
-        ledger_entry.debit_amount = ledger_entry.credit_amount = (
-            limit_fill_order.amount_to_sell.amount_decimal
-        )
-        ledger_entry.debit_conv = ledger_entry.credit_conv = limit_fill_order.conv
-        ledger_entry.cust_id = limit_fill_order.cust_id
-        ledger_entries.append(ledger_entry)
-    elif isinstance(limit_fill_order, FillOrder):
-        logger.info(
-            f"Fill order operation: {limit_fill_order.open_orderid} {limit_fill_order.current_owner}"
-        )
-        if not limit_fill_order.debit_conv or limit_fill_order.debit_conv.is_unset():
-            quote = await TrackedBaseModel.nearest_quote(timestamp=limit_fill_order.timestamp)
-            limit_fill_order.debit_conv = CryptoConv(
-                conv_from=limit_fill_order.open_pays.unit,  # HIVE
-                value=limit_fill_order.open_pays.amount_decimal,  # 25.052 HIVE
-                converted_value=limit_fill_order.current_pays.amount_decimal,  # 6.738 HBD
-                quote=quote,
-                timestamp=limit_fill_order.timestamp,
-            )
-        if not limit_fill_order.credit_conv or limit_fill_order.credit_conv.is_unset():
-            quote = await TrackedBaseModel.nearest_quote(timestamp=limit_fill_order.timestamp)
-            limit_fill_order.credit_conv = CryptoConv(
-                conv_from=limit_fill_order.current_pays.unit,  # HBD
-                value=limit_fill_order.current_pays.amount_decimal,  # 6.738 HBD
-                converted_value=limit_fill_order.open_pays.amount_decimal,  # 25.052 HIVE
-                quote=quote,
-                timestamp=limit_fill_order.timestamp,
-            )
-        seller_tracked = limit_fill_order.open_owner in tracked_customers
-        buyer_tracked = limit_fill_order.current_owner in tracked_customers
-
-        if buyer_tracked and seller_tracked:
-            # Both tracked: Create two entries as before
-            ledger_type = LedgerType.FILL_ORDER_BUY
-            buyer_entry = LedgerEntry(
-                group_id=f"{limit_fill_order.group_id}-{ledger_type}",
-                short_id=limit_fill_order.short_id,
-                timestamp=limit_fill_order.timestamp,
-                op_type=limit_fill_order.op_type,
-                ledger_type=ledger_type,
-                cust_id=limit_fill_order.cust_id,
-                description=f"Fill order buyer: {limit_fill_order.current_owner} pays {limit_fill_order.open_pays.amount_decimal} {limit_fill_order.open_pays.unit} for {limit_fill_order.current_pays.amount_decimal} {limit_fill_order.current_pays.unit}",
-                debit=AssetAccount(
-                    name="Customer Deposits Hive", sub=limit_fill_order.current_owner
-                ),  # Buyer debits their deposits for HIVE paid
-                debit_unit=limit_fill_order.open_pays.unit,  # HIVE
-                debit_amount=limit_fill_order.open_pays.amount_decimal,
-                debit_conv=limit_fill_order.debit_conv,
-                credit=AssetAccount(
-                    name="Escrow Hive", sub=limit_fill_order.current_owner
-                ),  # Buyer credits escrow for HBD received
-                credit_unit=limit_fill_order.current_pays.unit,  # HBD
-                credit_amount=limit_fill_order.current_pays.amount_decimal,
-                credit_conv=limit_fill_order.credit_conv,
-            )
-            ledger_entries.append(buyer_entry)
-
-            ledger_type = LedgerType.FILL_ORDER_SELL
-            seller_entry = LedgerEntry(
-                group_id=f"{limit_fill_order.group_id}-{ledger_type}",
-                short_id=limit_fill_order.short_id,
-                timestamp=limit_fill_order.timestamp,
-                op_type=limit_fill_order.op_type,
-                ledger_type=ledger_type,
-                cust_id=limit_fill_order.cust_id,
-                description=f"Fill order seller: {limit_fill_order.open_owner} receives {limit_fill_order.open_pays.amount_decimal} {limit_fill_order.open_pays.unit} for {limit_fill_order.current_pays.amount_decimal} {limit_fill_order.current_pays.unit}",
-                debit=AssetAccount(
-                    name="Escrow Hive", sub=limit_fill_order.open_owner
-                ),  # Seller debits escrow for HBD delivered
-                debit_unit=limit_fill_order.current_pays.unit,  # HBD
-                debit_amount=limit_fill_order.current_pays.amount_decimal,
-                debit_conv=limit_fill_order.credit_conv,  # Use credit_conv for seller's debit (HBD)
-                credit=AssetAccount(
-                    name="Customer Deposits Hive", sub=limit_fill_order.open_owner
-                ),  # Seller credits deposits for HIVE received
-                credit_unit=limit_fill_order.open_pays.unit,  # HIVE
-                credit_amount=limit_fill_order.open_pays.amount_decimal,
-                credit_conv=limit_fill_order.debit_conv,  # Use debit_conv for seller's credit (HIVE)
-            )
-            ledger_entries.append(seller_entry)
-        elif buyer_tracked and not seller_tracked:
-            # Buyer tracked, seller external: Single net entry for buyer, use suspense for external
-            ledger_type = LedgerType.FILL_ORDER_NET
-            net_entry = LedgerEntry(
-                group_id=f"{limit_fill_order.group_id}-{ledger_type}",
-                short_id=limit_fill_order.short_id,
-                timestamp=limit_fill_order.timestamp,
-                op_type=limit_fill_order.op_type,
-                ledger_type=ledger_type,
-                cust_id=limit_fill_order.cust_id,
-                description=f"Net fill order: {limit_fill_order.current_owner} trades {limit_fill_order.current_pays.amount_decimal} {limit_fill_order.current_pays.unit} for {limit_fill_order.open_pays.amount_decimal} {limit_fill_order.open_pays.unit} (external seller)",
-                debit=AssetAccount(
-                    name="Customer Deposits Hive", sub=limit_fill_order.current_owner
-                ),  # Buyer debits deposits for HIVE paid
-                debit_unit=limit_fill_order.open_pays.unit,
-                debit_amount=limit_fill_order.open_pays.amount_decimal,
-                debit_conv=limit_fill_order.debit_conv,
-                credit=AssetAccount(
-                    name="Customer Deposits Hive", sub=limit_fill_order.current_owner
-                ),  # Buyer credits deposits for HBD received (net effect)
-                credit_unit=limit_fill_order.current_pays.unit,
-                credit_amount=limit_fill_order.current_pays.amount_decimal,
-                credit_conv=limit_fill_order.credit_conv,
-            )
-            # Optionally, add a suspense entry for the external side (but don't save it if netting)
-            # suspense_entry = LedgerEntry(... LiabilityAccount("External Market Suspense", sub="untracked") ...)
-            ledger_entries.append(net_entry)
-        else:
-            # Neither tracked: Skip or log (not relevant to your entity)
-            logger.info(
-                f"Fill order between untracked parties: {limit_fill_order.current_owner} and {limit_fill_order.open_owner}. Skipping."
-            )
-            return []
-
-    # Save all entries
-    for entry in ledger_entries:
-        await entry.save()
-
-    return ledger_entries  # Return the list
-
-
-# MARK: CustomJson Operations
-async def process_custom_json(
-    custom_json: CustomJson, nobroadcast: bool = False
-) -> List[LedgerEntry]:
-    """
-    Processes a CustomJson operation and creates a ledger entry if applicable.
-
-    This method handles CustomJson operations, ensuring that appropriate debit and credit
-    accounts are assigned based on the operation type. If a ledger entry with the same group_id
-    already exists, the operation is skipped.
-
-    Returns:
-        LedgerEntry: The created or existing ledger entry, or None if no entry is created.
-    """
-    server_id = InternalConfig().server_id
-    if custom_json.cj_id in ["v4vapp_notification", "v4vapp_dev_notification"]:
-        logger.info(f"Notification CustomJson: {custom_json.json_data.memo}")
-        return None
-    if custom_json.cj_id in ["v4vapp_dev_transfer", "v4vapp_transfer"]:
-        keepsats_transfer = KeepsatsTransfer.model_validate(custom_json.json_data)
-        keepsats_transfer.msats = (
-            (keepsats_transfer.sats * 1000)
-            if keepsats_transfer.sats and not keepsats_transfer.msats
-            else keepsats_transfer.msats
-        )
-        # MARK: CustomJson Transfer user to user
-        if (
-            custom_json.from_account
-            and custom_json.to_account
-            and keepsats_transfer.msats
-            and custom_json.from_account != custom_json.to_account
-        ):
-            if not custom_json.conv or custom_json.conv.is_unset():
-                quote = await TrackedBaseModel.nearest_quote(timestamp=custom_json.timestamp)
-                await custom_json.update_conv(quote=quote)
-            ledger_entries = await custom_json_internal_transfer(
-                custom_json=custom_json, keepsats_transfer=keepsats_transfer
-            )
-            # Check for a parent id to see if this is a reply transaction
-            if keepsats_transfer.parent_id:
-                # This is a reply transaction, we need to process it as such
-
-                parent_op = await load_tracked_object(tracked_obj=keepsats_transfer.parent_id)
-                if parent_op:
-                    parent_op.add_reply(
-                        reply_id=custom_json.group_id_p,
-                        reply_type=ReplyType.CUSTOM_JSON,
-                        reply_msat=keepsats_transfer.msats if keepsats_transfer.msats else 0,
-                        reply_message="Reply to transfer",
-                    )
-                    await parent_op.save()
-                    # Now we may need to process if the parent op is an Invoice we need to send the keepsats
-                    # to the correct destination.
-                    if isinstance(parent_op, Invoice) and custom_json.from_account == server_id:
-                        await process_lightning_receipt_stage_2(
-                            invoice=parent_op, nobroadcast=nobroadcast
-                        )
-                        return ledger_entries
-                    if isinstance(parent_op, CustomJson) and custom_json.to_account == server_id:
-                        # Process this as if it were a request to convert Keepsats to Hive/HBD
-                        logger.info(
-                            f"Processing Keepsats to Hive conversion: {custom_json.json_data.memo}"
-                        )
-                        # await conversion_keepsats_to_hive(server_id= server_id, cust_id=tracked_op=custom_json, nobroadcast=nobroadcast)
-
-            if custom_json.to_account == server_id:
-                # Process this as if it were an inbound Hive transfer with a memo.
-                await follow_on_transfer(tracked_op=custom_json, nobroadcast=nobroadcast)
-
-            return []
-
-        # MARK: CustomJson to pay a lightning invoice
-        # If this has a memo that should contain the invoice and the instructions like "#clean"
-        # invoice_message we will use to send on if we generate an invoice form a lightning address
-        elif keepsats_transfer.memo and keepsats_transfer.to_account == server_id:
-            # This is a transfer operation, we need to process it as such
-
-            if not custom_json.conv or custom_json.conv.is_unset():
-                await custom_json.update_conv()
-                if custom_json.conv and custom_json.conv.is_unset():
-                    raise LedgerEntryCreationException(
-                        "Conversion not set in CustomJson operation."
-                    )
-
-            await follow_on_transfer(tracked_op=custom_json, nobroadcast=nobroadcast)
-            return []
-
-    logger.error(
-        f"CustomJson operation not implemented for v4vapp_group_id: {custom_json.group_id}.",
-        extra={"notification": False, **custom_json.log_extra},
-    )
-    raise NotImplementedError(
-        f"Some other custom_json functionality which hasn't been implemented yet {custom_json.group_id}."
-    )
